% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_betas.R
\name{add_kappas}
\alias{add_kappas}
\title{Calculate force of infection under different contact rates, isolation
measures and social distancing scenarios.}
\usage{
add_kappas(pars, mult = FALSE)
}
\arguments{
\item{pars}{A parameter object as returned by \code{get_parameters}.}

\item{mult}{A logical indicating whether social distancing and isolation
measures have a combined effect.}
}
\description{
Calculate force of infection under different contact rates, isolation
measures and social distancing scenarios.
}
