% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rates.R
\name{add_rates}
\alias{add_rates}
\title{Calculate transition rates from periods.}
\usage{
add_rates(pars, state = NULL)
}
\arguments{
\item{pars}{A parameter object as returned by \code{get_parameters}.}

\item{state}{An optional state of compartments, used to adjust rates for
hospital occupancy.}
}
\description{
Calculate transition rates from periods.
}
