% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_to_ifr.R
\name{age_to_ifr}
\alias{age_to_ifr}
\title{Calculate IFR from ages using meta-analysis from Levin et al (2020).}
\source{
Levin et al. (2020). https://link.springer.com/article/10.1007/s10654-020-00698-1#Sec7
}
\usage{
age_to_ifr(age)
}
\arguments{
\item{age}{A vector of ages.}
}
\description{
Calculate IFR from ages using meta-analysis from Levin et al (2020).
}
