% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simex.R
\name{as.simex}
\alias{as.simex}
\title{shape dust2 simulation outputs or monty trajectory samples into the
simex format. dimensions names must be specified in dims, dimension
values can be optionally passed via ...}
\usage{
as.simex(x, sys, pars, dims = c("state", "time"), ...)
}
\description{
shape dust2 simulation outputs or monty trajectory samples into the
simex format. dimensions names must be specified in dims, dimension
values can be optionally passed via ...
}
