% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.simex.R, R/extract.simex_defunct.R
\name{extract}
\alias{extract}
\title{Define generic extract method.}
\usage{
extract(x, ...)

extract(x, ...)
}
\description{
Define generic extract method.

Define generic extract method.
}
