% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.simex.R
\name{extract.simex}
\alias{extract.simex}
\title{Extract formatted dataframes from simex object.}
\usage{
\method{extract}{simex}(
  simex,
  what = c("prevalence", "incidence"),
  cri = FALSE,
  cri_alpha = 0.95,
  filter = NULL,
  stratify_by = c("time", "vax", "compartment", "age")
)
}
\arguments{
\item{simex}{A simex object returned by \code{run_model}.}

\item{what}{What to plot: one of "prevalence" (number in each category per
day), "deltas" (change in prevalence from one day to the next) and
"incidence" (new additions to that compartment per day).}

\item{stratify_by}{Variables to stratify output but, must be one or more of
"time", "age", "compartment", "vax".}
}
\description{
Extract formatted dataframes from simex object.
}
