% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.simex_defunct.R
\name{extract2.simex}
\alias{extract2.simex}
\title{Extract formatted dataframes from simex object.}
\usage{
extract2.simex(
  simex,
  what = c("prevalence", "deltas", "incidence"),
  stratify_by = c("day", "vax", "compartment")
)
}
\arguments{
\item{simex}{A simex object returned by \code{run_model}.}

\item{what}{What to plot: one of "prevalence" (number in each category per
day), "deltas" (change in prevalence from one day to the next) and
"incidence" (new additions to that compartment per day).}

\item{stratify_by}{Variables to stratify output but, must be one or more of
"day", "age", "compartment", "vax".}
}
\description{
Extract formatted dataframes from simex object.
}
