% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{extract_active_par}
\alias{extract_active_par}
\title{Extractive active values of a given parameter name}
\usage{
extract_active_par(input, name, active_par)
}
\description{
Extractive active values of a given parameter name
}
