% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_html_list}
\alias{format_html_list}
\title{Format character vector into HTML bulleted list}
\usage{
format_html_list(char, ordered = FALSE)
}
\arguments{
\item{char}{a character vector. Each element will be a bullet}

\item{ordered}{logical (T/F). If \code{TRUE}, return numbered list.}
}
\description{
Format character vector into HTML bulleted list
}
