% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrices.R
\name{get_agesums}
\alias{get_agesums}
\title{The sum of "ages" of infections caused by each compartment in each
compartment.}
\usage{
get_agesums(p_trans, pars)
}
\arguments{
\item{p_trans}{Probability of infection per contacts.}

\item{pars}{A set of parameters as returned by \code{get_parameters}.}
}
\description{
The sum of "ages" of infections caused by each compartment in each
compartment.
}
