% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arg_table.R
\name{get_default}
\alias{get_default}
\title{Extract default value call as string from formal.}
\usage{
get_default(arg, what = c("text", "value"))
}
\arguments{
\item{arg}{An argument returned by \code{formals}.}

\item{what}{Return deparsed text (\code{"text"}) or evaluated value
(\code{"value"}).}
}
\description{
Extract default value call as string from formal.
}
