% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arg_table.R
\name{get_description}
\alias{get_description}
\title{Extract description of function argument from documentation.}
\usage{
get_description(fn_name, arg_name)
}
\arguments{
\item{fn_name}{A string describing the function name. This must be loaded
into the namespace.}

\item{arg_name}{A string describing the argument of the function.}
}
\description{
Extract description of function argument from documentation.
}
