% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrices.R
\name{get_doubling_time}
\alias{get_doubling_time}
\title{The doubling is calculated from the exponential growth rate.}
\usage{
get_doubling_time(p_trans, pars)
}
\arguments{
\item{p_trans}{Probability of infection per contacts.}

\item{pars}{A set of parameters as returned by \code{get_parameters}.}
}
\description{
The doubling is calculated from the exponential growth rate.
}
