% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrices.R
\name{get_generation_time}
\alias{get_generation_time}
\title{The mean generation time is calculated as the sum of "ages" of cases caused
divided by the number of cases caused.}
\usage{
get_generation_time(p_trans, pars)
}
\arguments{
\item{p_trans}{Probability of infection per contacts.}

\item{pars}{A set of parameters as returned by \code{get_parameters}.}
}
\description{
The mean generation time is calculated as the sum of "ages" of cases caused
divided by the number of cases caused.
}
