% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrices.R
\name{get_matrices}
\alias{get_matrices}
\title{Calculate next-gen (and other) matrices used for R0 and r calculations: these
assume a completely susceptible population and no interventions!}
\usage{
get_matrices(p_trans, pars)
}
\arguments{
\item{p_trans}{Probability of infection per contacts.}

\item{pars}{A set of parameters as returned by \code{get_parameters}.}
}
\description{
Calculate next-gen (and other) matrices used for R0 and r calculations: these
assume a completely susceptible population and no interventions!
}
