% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priors.R
\name{get_priors}
\alias{get_priors}
\title{Create a Monty Prior Model from List Specifications}
\usage{
get_priors(...)
}
\arguments{
\item{...}{Named lists. The name corresponds to the parameter name.
Each list must contain a \code{dist} element (case-insensitive)
and parameters matching a signature in \code{monty_dsl_distributions()}.}
}
\value{
A \code{monty_model} object.
}
\description{
A wrapper around \code{monty::monty_dsl} that dynamically constructs
prior distributions using list inputs, validated against \code{monty_dsl_distributions()}.
}
