% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrices.R
\name{get_r}
\alias{get_r}
\title{Calculate the exponential growth rate as the principal eigenvalue of the
change in exposed.}
\usage{
get_r(p_trans, pars)
}
\arguments{
\item{p_trans}{Probability of infection per contacts.}

\item{pars}{A set of parameters as returned by \code{get_parameters}.}
}
\description{
Calculate the exponential growth rate as the principal eigenvalue of the
change in exposed.
}
