% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrices.R
\name{get_R0}
\alias{get_R0}
\title{Calculate R0 as the principal eigenvalue of the next-generation matrix.}
\usage{
get_R0(p_trans, pars)
}
\arguments{
\item{p_trans}{Probability of infection per contacts.}

\item{pars}{A set of parameters as returned by \code{get_parameters}.}
}
\description{
Calculate R0 as the principal eigenvalue of the next-generation matrix.
}
