% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_model.R
\name{ode_model}
\alias{ode_model}
\title{The series of ODEs passed to deSolve.}
\usage{
ode_model(time, state, pars)
}
\arguments{
\item{time}{A vector of days.}

\item{state}{A matrix indicating the state of number of individuals in every
compartment. Rows represent age cateogorie, columns the disease
compartment.}

\item{pars}{A parameter set as returned by \code{get_parameters}.}
}
\description{
The series of ODEs passed to deSolve.
}
\author{
Finlay Campbell, Prabasaj Paul
}
