\name{Overview}
\alias{Overview}
\title{Overview over Package simex}
\description{Package \code{simex} is an implementation of the SIMEX--algorithm by Cook and Stephanski and the MCSIMEX--Algorithm by Küchenhof, Mwalili and Lesaffre.
}
\details{
The package includes first of all the implementation for the SIMEX-- and MCSIMEX--Algorithms. Jackknife and asymptotic variance estimation is implemented.
Various methods and analytic tools are provided for a simple and fast access to the SIMEX-- and MCSIMEX--Algorithm.
}
\seealso{\code{\link{simex}},\code{\link{mcsimex}},\code{\link{refit}}}
\keyword{models}

