% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{shiny_to_simex}
\alias{shiny_to_simex}
\title{Shape shiny parameter to fit simex model input}
\usage{
shiny_to_simex(input, active_par)
}
\description{
Shape shiny parameter to fit simex model input
}
