% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{simex_to_shiny}
\alias{simex_to_shiny}
\title{Generate shiny output for a given parameter tab}
\usage{
simex_to_shiny(simex_input, tab_id, simex_defaults)
}
\description{
Generate shiny output for a given parameter tab
}
