% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simex.R
\name{simulate_from_samples}
\alias{simulate_from_samples}
\title{Simulate forward from a snapshot}
\usage{
simulate_from_samples(
  gen,
  samples,
  packer,
  time = 0:100,
  start_from_snapshot = NULL,
  modification = NULL
)
}
\arguments{
\item{start_from_snapshot}{If not NULL, an integer indicating which
snapshot to use as starting state.}

\item{start_time}{The model time the simulation starts from}
}
\description{
Simulate forward from a snapshot
}
