% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ode.R
\name{solve_ode}
\alias{solve_ode}
\title{Solve the ODE given a starting state and specific set of parameters.}
\usage{
solve_ode(pars, days, state)
}
\arguments{
\item{pars}{A single parameter set as returned by \code{get_parameters}.}

\item{days}{A vector of days to solve the ODE across.}

\item{state}{A matrix indicating the state of number of individuals in every
compartment. Rows represent age cateogorie, columns the disease
compartment.}
}
\description{
Solve the ODE given a starting state and specific set of parameters.
}
\author{
Finlay Campbell, Prabasaj Paul
}
