% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_comparison_defunct.R
\name{vis_comparison2}
\alias{vis_comparison2}
\title{Compare different runs.}
\usage{
vis_comparison2(
  simexl,
  what = c("prevalence", "deltas", "incidence"),
  format = c("timeline", "endpoint"),
  type = c("ggplot", "highchart"),
  log = FALSE,
  use_absolute_numbers = FALSE,
  show_compartment = c("D", "S", "E", "C", "H", "R"),
  freescales = TRUE,
  base_size = 11
)
}
\arguments{
\item{simexl}{The solved system of equations as returned by \code{run_model}.}

\item{what}{What to plot: one of "prevalence" (number in each category per
day), "deltas" (change in prevalence from one day to the next) and
"incidence" (new additions to that compartment per day).}

\item{format}{One of "timeline" (shows a timeline stratified by compartment
and vaccination status) or "endpoint" (shows the final value stratified by
age category).}

\item{log}{Logical indicating whether a log scale should be used.}

\item{use_absolute_numbers}{Logical indicating whether absolute absolute
numbers (in terms of individuals) should be shown or relative to population
size.}

\item{show_compartment}{The compartment to show when using \code{format =
"endpoint"}.}

\item{freescales}{Logical indicating whether the different comparments should
use different y-axes.}

\item{base_size}{Base size passed to theme_*}
}
\description{
Compare different runs.
}
\author{
Finlay Campbell
}
