% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.simex.R
\name{extract}
\alias{extract}
\title{Generic extract method for simex objects}
\usage{
extract(x, ...)
}
\arguments{
\item{x}{Object to extract from.}

\item{...}{Further arguments passed to methods.}
}
\description{
Dispatches to \code{\link[=extract.simex]{extract.simex()}} when \code{x} inherits from \code{"simex"}.
}
