% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arg_table.R
\name{get_arg_table}
\alias{get_arg_table}
\title{Generate table describing function arguments and their defauts.}
\usage{
get_arg_table(fn_name)
}
\arguments{
\item{fn_name}{A string describing the function name. This must be loaded
into the namespace.}
}
\description{
Generate table describing function arguments and their defauts.
}
