% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simex.R
\name{plot.simex}
\alias{plot.simex}
\title{Plot timeline of prevalence or incidence.}
\usage{
\method{plot}{simex}(
  simex,
  what = c("prevalence", "deltas", "incidence"),
  log = FALSE,
  freescales = TRUE
)
}
\arguments{
\item{simex}{The solved system of equations as returned by \code{run_model}.}

\item{what}{What to plot: one of "prevalence" (number in each category per
day), "deltas" (change in prevalence from one day to the next) and
"incidence" (new additions to that compartment per day).}

\item{log}{Logical indicating whether a log scale should be used.}

\item{freescales}{Logical indicating whether the different comparments should
use different y-axes.}
}
\description{
Plot timeline of prevalence or incidence.
}
\author{
Finlay Campbell
}
