% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simex.R
\name{plot.simex}
\alias{plot.simex}
\title{Plot timeline of prevalence or incidence.}
\usage{
\method{plot}{simex}(
  simex,
  what = c("prevalence", "deltas", "incidence"),
  format = c("timeline", "endpoint"),
  log = FALSE,
  freescales = TRUE,
  show_hosp_capacity = TRUE,
  use_absolute_numbers = FALSE,
  show_compartment = c("D", "S", "E", "C", "H", "R"),
  base_size = 11,
  type = c("ggplot", "highchart")
)
}
\arguments{
\item{simex}{The solved system of equations as returned by \code{run_model}.}

\item{what}{What to plot: one of "prevalence" (number in each category per
day), "deltas" (change in prevalence from one day to the next) and
"incidence" (new additions to that compartment per day).}

\item{format}{One of "timeline" (shows a timeline stratified by compartment
and vaccination status) or "endpoint" (shows the final value stratified by
age category).}

\item{log}{Logical indicating whether a log scale should be used.}

\item{freescales}{Logical indicating whether the different comparments should
use different y-axes.}

\item{show_hosp_capacity}{Logical indicating whether hospital capacity should be
displayed in the "H" compartment (only possible when what = "prevalence").}

\item{use_absolute_numbers}{Logical indicating whether absolute absolute
numbers (in terms of individuals) should be shown or relative to population
size.}

\item{show_compartment}{The compartment to show when using \code{format = "endpoint"}.}

\item{base_size}{Base size passed to theme_*.}

\item{type}{"ggplot" will return a static ggplot figure, "highchart" will
return a dynamic figure.}
}
\description{
Plot timeline of prevalence or incidence.
}
\author{
Finlay Campbell
}
