\name{print.SIMEX}
\alias{print.SIMEX}
\alias{print.MCSIMEX}
\title{Print method for the SIMEX and MCSIMEX class}
\description{Printing the most important values in a clear way}
\usage{
print.SIMEX(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
  \item{x}{object of class \code{SIMEX}}
  \item{digits}{number of digits to be printed}
  \item{\dots}{arguments passed to other functions}
}
\author{Wolfgang Lederer}
\keyword{models}
