% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simex.R
\name{print.simex}
\alias{print.simex}
\title{Printing method for simex object.}
\usage{
\method{print}{simex}(simex)
}
\arguments{
\item{simex}{A simex object returned by \code{run_model}.}
}
\description{
Printing method for simex object.
}
