% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the model on on a single set of parameters, or a list of parameters.}
\usage{
run_model(pars, init_state = NULL, max_day = 365)
}
\arguments{
\item{pars}{A parameter set as returned by \code{get_parameters}, or a named
list of parameters where the name indicates the start date of the given
parameter set.}

\item{max_day}{}
}
\description{
Run the model on on a single set of parameters, or a list of parameters.
}
\author{
Finlay Campbell, Prabasaj Paul
}
