% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simex-package.R
\docType{package}
\name{simex-package}
\alias{simex}
\alias{simex-package}
\title{simex: Simulating epidemics and interventions}
\description{
Simulate epidemics using country-specific contact rates.
}
\author{
\strong{Maintainer}: Finlay Campbell \email{campellf@who.int}

}
\keyword{internal}
