% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simex-package.R
\docType{package}
\name{simex-package}
\alias{simex}
\alias{simex-package}
\title{simex: Simulating epidemics and interventions}
\description{
Simulate epidemics and interventions using country-specific contact rates and age demographics.
}
\author{
\strong{Maintainer}: Finlay Campbell \email{campbellf@who.int}

Authors:
\itemize{
  \item Prabasaj Paul \email{ppaul@who.int}
}

}
\keyword{internal}
